Blendshape tangets & normals, cleanup.
Turns out Maya was always including normals in the FBX export, they were
just a bit trickier to get to than originally surmised. We need to go
through the proper element access formalities that takes mapping and
reference modes into account.

Luckily we already have a helper class for this, so let's lean on that.