Mocha integration test suite (#150)
This moves ./npm to ./npm/fbx2gltf, and introduces a Mocha / TypeScript based testing harness. 

We're also adding a GIT LFS dependency, but only for people who want to run tests. We're using LFS to store the many FBX models we want to convert to GLB as part of the integration test suite.